+++ PlantUmlViewer/PlantUmlViewer/DiagramGeneration/PlantUml/ErrorFormat.cs
@@ -1,0 +1,9 @@
+﻿namespace PlantUmlViewer.DiagramGeneration.PlantUml
+{
+    public enum ErrorFormat
+    {
+        TwoLines,
+        SingleLine,
+        Verbose
+    }
+}