+++ PlantUmlViewer/PlantUmlViewer/DiagramGeneration/PlantUml/OutputFormat.cs
@@ -1,0 +1,17 @@
+﻿namespace PlantUmlViewer.DiagramGeneration.PlantUml
+{
+    public enum OutputFormat
+    {
+        Png,
+        Svg,
+        Eps,
+        Pdf,
+        Vdx,
+        Xmi,
+        Scxml,
+        Html,
+        Ascii,
+        AsciiUnicode,
+        LaTeX
+    }
+}