+++ PlantUmlViewer/PlantUmlViewer/Forms/ControlsExtensions.cs
@@ -5,7 +5,7 @@ using System.Windows.Forms;
 
 namespace PlantUmlViewer.Forms
 {
-    public static class ControlsExtensions
+    internal static class ControlsExtensions
     {
         public static void InvokeIfRequired(this Control control, Action action)
         {